Move Initial topography from the main directory to geometry model and also put it into that subsection.
