+++ include/aspect/geometry_model/initial_topography_model/interface.h
@@ -19,8 +19,8 @@
 */
 
 
-#ifndef __aspect__initial_topography_model_interface_h
-#define __aspect__initial_topography_model_interface_h
+#ifndef __aspect__geometry_model__initial_topography_model_interface_h
+#define __aspect__geometry_model__initial_topography_model_interface_h
 
 #include <aspect/plugins.h>
 #include <deal.II/base/parameter_handler.h>
