+++ include/aspect/geometry_model/initial_topography_model/zero_topography.h
@@ -19,10 +19,10 @@
 */
 
 
-#ifndef __aspect__initial_topography_model_zero_topography_h
-#define __aspect__initial_topography_model_zero_topography_h
+#ifndef __aspect__geometry_model__initial_topography_model_zero_topography_h
+#define __aspect__geometry_model__initial_topography_model_zero_topography_h
 
-#include <aspect/initial_topography_model/interface.h>
+#include <aspect/geometry_model/initial_topography_model/interface.h>
 
 namespace aspect
 {
